[fluorine] Merge forward from 2018.3 to fluorine

+++ tests/support/gitfs.py
@@ -417,6 +417,11 @@ class GitPillarTestBase(GitTestBase, LoaderModuleMockMixin):
         with salt.utils.files.fopen(
                 os.path.join(self.admin_repo, 'bar.sls'), 'w') as fp_:
             fp_.write('included_pillar: True\n')
+        # Add another file in subdir
+        os.mkdir(os.path.join(self.admin_repo, 'subdir'))
+        with salt.utils.files.fopen(
+                os.path.join(self.admin_repo, 'subdir', 'bar.sls'), 'w') as fp_:
+            fp_.write('from_subdir: True\n')
         _push('master', 'initial commit')
 
         # Do the same with different values for "dev" branch
@@ -462,7 +467,7 @@ class GitPillarTestBase(GitTestBase, LoaderModuleMockMixin):
         # The top.sls should be the only file in this branch
         self.run_function(
             'git.rm',
-            [self.admin_repo, 'foo.sls'],
+            [self.admin_repo, 'foo.sls', os.path.join('subdir', 'bar.sls')],
             user=user)
         with salt.utils.files.fopen(
                 os.path.join(self.admin_repo, 'top.sls'), 'w') as fp_:
@@ -472,6 +477,23 @@ class GitPillarTestBase(GitTestBase, LoaderModuleMockMixin):
                 - bar
             '''))
         _push('top_only', 'add top_only branch')
+
+        # Create just another top file in a separate repo, to be mapped to the base
+        # env and including mounted.bar
+        self.run_function(
+            'git.checkout',
+            [self.admin_repo],
+            user=user,
+            opts='-b top_mounted')
+        # The top.sls should be the only file in this branch
+        with salt.utils.files.fopen(
+                os.path.join(self.admin_repo, 'top.sls'), 'w') as fp_:
+            fp_.write(textwrap.dedent('''\
+            base:
+              '*':
+                - mounted.bar
+            '''))
+        _push('top_mounted', 'add top_mounted branch')
 
     def make_extra_repo(self, root_dir, user='root'):
         self.bare_extra_repo = os.path.join(root_dir, 'extra_repo.git')
@@ -538,6 +560,8 @@ class GitPillarSSHTestBase(GitPillarTestBase, SSHDMixin):
 
     @classmethod
     def tearDownClass(cls):
+        if cls.case is None:
+            return
         if cls.case.sshd_proc is not None:
             cls.case.sshd_proc.send_signal(signal.SIGTERM)
         cls.case.run_state('user.absent', name=cls.username, purge=True)
